+++ package.json
@@ -39,6 +39,7 @@
 		"check.tsc": "tsc --incremental --noEmit"
 	},
 	"dependencies": {
+		"@tabler/icons-react": "^3.5.0",
 		"clsx": "^2.1.1",
 		"react": "^18.3.1",
 		"react-dom": "^18.3.1"
