+++ src/form/Input.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 import { Input } from './Input';
 
 export default {
-	title: 'Input',
+	title: 'form/Input',
 	component: Input,
 } as Meta<typeof Input>;
